fix Imported names order warning with google style
using the following import:

```python
from .models import (
    A,
    C,
    b,
)
```

I would get this message:

```
I101 Imported names are in the wrong order. Should be A, C, b
```

This happened because the comparison that the plugin does is
correct (lexicographic, case-insensitive). But when the error message is
displayed, only a lexicographic case-sensitive sorting algorithm is
used (python's default sorted)