Refactor structure
Flake8-Import-Order was originally written for a single style, namely
cryptography over time the google, smarkets and most recently pep8
styles have been added. These additions patched the styles into the
codebase, whereas this commit refactors the codebase to make differing
styles the norm.

This introduces a potentially breaking change to the cryptography
style, which is best seen in the complete_with_bad_order.py test
case. In this case Flake8-Import-Order would previously insist that
`from Z.A.B import A` be before `from Z import A` whereas now `from
Z.A import A` must be before `from Z.A.B import A`. This I believe is
the expected cryptography style.

Finally this drops support for Python2.6, to allow for set and dict
comprehension usage.