Correct ungrouped I100 reporting case sensitivity
Previously checking of ungrouped imports would take case into account
such that

    from . import a
    from . import B

would be reported as an I100 error. This is fixed such that the
ordering follows the styles' name ordering. This affects Google,
Smarkets and AppNexus styles.

Note that the ordering of the import names is same as if the imported
names were grouped.