Refactor test structure
This refactors the tests into a set of test cases for the various
styles alongside separate tests for the flake8 and pylama
linters. This includes additional test cases for lexicographic
ordering and removes tests that are no longer required.

This structure should be quicker to run given the change from writing
files to using in memory (stdin) based checks for the stdlib tests.