Enable application-package-names for pep8 style
For networking-midonet project, [1] we want to use pep8 style
with an additional group of modules. [2]
application-package-names seems to fit our needs well.  I think
it can be used for the rest of OpenStack projects as well.

While this can be a dedicated style, I don't think it's worth
doing it, given that it's compatible with the existing behaviour
as far as application-package-names is left to be empty, which
is the default.

[1] https://docs.openstack.org/networking-midonet/latest/
[2] https://review.openstack.org/#/c/482450/1/tox.ini@86