Fix cryptography same_section check
The cryptography first party (application imports) should be grouped
together by packages and with all the relative imports grouped
together. The previous version would not correctly enforce this, due
to a poorly definition of relative imports (including non-relative
first party). The standard style same section check is also updated to
make it clear that it is both first party, not both relative that
matters.