Fix I202 bug with multiline imports
Previously any multiline import would be considered as taking up a
single line. This would cause the linter to assume there are
additional newlines and hence raise an I202 error. For example

    from X import A, a, \
        B, b
    from Y import A

would raise an I202 on the 3rd line.

To fix this it is required to know the line of the final token in the
import. The final token is key due to paranthesis for example,

    from X import (
        A, a, B, b,
    )
    from Y import A

the first three lines are all the same import.

As the stdlib ast module cannot provide the final token location, the
asttokens package is used.