Fix no invocation bug and asttokens usage
Without a `tree` argument to the flake8 linter the plugin is not
invoked and silently does nothing, which indicates to the user that
the linting passed (although nothing happened).

Whilst fixing this I noticed the asttokens usage was a bit rushed and
not only was the ast tree useful, the asttokens marked tree was
created in an unclear way (is it really part of the load file?). This
was also fixed.