Fix noqa acceptance
As the I202 error requires knowledge of the previous import to work it
is necessary not to remove the noqa lines before checking the style,
and instead check after (as this patch does). This should remove errors
such as,

    import A
    # some comment
    import B
    import C

the 3rd line will trigger a I202 error due to the comment, which can
be ignored. However with the previous version this would cause line 4
to also raise a I202 as only import A and import C will be style
checked.