Change determination of spacing
This ensures that a newline space is only considered to be a blank
line i.e. newline. It prevents things like comments or a try/except
triggering a spacing error (or not triggering for I201).

It does so by considering import and newline nodes, which is a change
to how the style checkers work and hence it can break custom styles if
they consider the nodes directly (most do not).

It allows asttokens to be dropped as a dependency as the end of the
statement is no longer required (as the checker knows where the newlines
are instead).