Understand namespace packages
This change means that the root package name is never considered to a
namespace package, but rather the root package. This shouldn't make
any difference in normal usage, however it allows for one namespaced
package to be local and another third party.

This change only works in Python3 as I'm unsure how to find a module's
spec in Python2. A Python2 compatible method of importing the package
and testing for the namespace properties was rejected due to the
likely performance cost and due to concerns about import side
affects. I don't intend to support Python2 for many more releases so
this is an acceptable compromise.