Add remaining python3 modules
I've noted a couple missing packages, so today I wanted to see if I could get the remainder. I have the html version of the 3.6.3 docs downloaded locally for quick reference, so I figured this would be an easy to scrape most of the missing modules.

I wrote a quick script to import the scraped modules in a fresh venv to make sure all of these are importable and hopefully eliminate any false positives. I used a combination of the below strategies, all of which were run from `python-3.6.3-docs-html/library`

Collect the HTML filenames, which for the most part seem to be importable modules, and if they are importable and not in `stdlib.txt` (a text file of flake8-import-order's existing stdlib list), then accumulate them:
```shell_session
for htmlfile in *.html; do lib="${htmlfile%.html}"; if ! grep --quiet "${lib}" stdlib.txt; then python3 -c "import ${lib}" &> /dev/null && echo "${lib}" >> good_imports.txt; fi; done
```

Scrape the HTML files for any `<div>` with an `id` starting with `module-`:
```shell_session
$ grep -h -P -o '(?<=<div class="section" id="module-).*(?=">)' *.html | sort -u > all_modules.txt
```

Try to import each module; accumulate the ones that work and print the ones that result in an error.
```shell_session
$ for lib in $(cat all_modules.txt); do python3 -c "import ${lib}" &> /dev/null && echo "${lib}" >> good_imports.txt || echo "${lib}"; done
```

Then I just basically did a `sort --unique` with the existing stdlib list.

Anyway, may want to look it over to make sure I didn't screw something up, but I think that having tested for import-ability in a fresh venv should mean all the additions are valid.