Improve the error messages and typing
This hopefully makes the errors clearer to the user, thereby allowing
them to better understand how to fix the error. As it explains the
sections determined and will note if the section is wrong when
considering the relative ordering.

This also uses a Enum for convenience, which in turn improves the type
usage.