Lazily build grouping explanation string
The previous logic would build the grouping explanation string for every
import, even if there isn't an error condition. This was unnecessary, so
we now defer that formatting work to a static method that is only
invoked when an error is detected.

Also, change the second `if` to an `elif` because these cases can't both
be true at the same time.